♻️ refactor: improve auth flow and navigation structure
The commit reorganizes the authentication and navigation flow by:
1. Moving auth redirects from layout to individual screens
2. Simplifying navigation structure with proper initial routes
3. Removing redundant subscription guard checks
