+++ apps/expo/src/app/(auth)/_layout.tsx
@@ -1,21 +1,5 @@
-import { Redirect, Stack } from "expo-router";
-import { useAuth } from "@clerk/clerk-expo";
-
-import { useAppStore } from "~/store";
+import { Stack } from "expo-router";
 
 export default function AuthRoutesLayout() {
-  const { isSignedIn } = useAuth();
-  const hasCompletedOnboarding = useAppStore(
-    (state) => state.hasCompletedOnboarding,
-  );
-
-  if (isSignedIn && hasCompletedOnboarding) {
-    return <Redirect href={"/feed"} />;
-  }
-
-  if (isSignedIn && !hasCompletedOnboarding) {
-    return <Redirect href={"/onboarding"} />;
-  }
-
   return <Stack />;
 }
